python/lc-stats/pyproject.toml
[project]
name = "lc-stats"
version = "0.1.0"
description = "LeetCode progress statistics generator with visualizations"
readme = "README.md"
authors = [
    { name = "Leo", email = "liaofeng0203@gmail.com" }
]
requires-python = ">=3.12"
dependencies = [
    "pyyaml>=6.0",
    "matplotlib>=3.8",
    "wordcloud>=1.9",
    "pillow>=10.0",
]

[project.scripts]
lc-stats = "lc_stats:main"

[build-system]
requires = ["uv_build>=0.8.22,<0.9.0"]
build-backend = "uv_build"
